Finally attempt to make the dinosaur jump
I say attempt, because I failed on this attempt. From what I can see
from the debug data, the sensors are reading the individual arms of the
cactus as separate obstacles, and that's throwing off all the
calculations for identifying obstacles. I'm sure there's a way to fix
this, there's always a way. I just need to find out how.

I really home I don't have to go back to the running mean approach from
my very first prototype. That will be very resource intensive,
especially now that there are four photoresistors being polled in
parallel.

Maybe what I'll do is determine a safe amount of time to wait so that
the cactus is probably still the same color so that the arduino still
recognizes it.

Or I could use optics to blur the light coming into the photoresistor
(using something diffuse like hot-melt glue, which I have an abundance
of). I know commit messages are meant more for explaining the changes in
the code, but my brain is really on its last leg tonight and I really
need a reliable way to write notes to myself about this code. This is
the way.