Get the dinosaur to actually jump
It's a little inconsistent with timing, however. My plan is to redesign
the physical sensor array and make it a little bit wider, to reduce the
amount of jitter in velocity calculations.

HOWEVER, first I need to figure out a way to use timer interrupts to
check the Obstacle array.