Fix user I/O bugs and add stop button
Among these were weird line spacing issues when reading the serial
monitor through the python script, and the menu permanently freezing at
seemingly random times when waiting for user input.

The good news is that there is finally a reliable stop button that is
active whenever the game is running (as opposed to menu stuff). That
way, if the Arduino fails to play the game, you can always exit to the
main menu and do some other stuff.

Next up is adding actual configuration functionality accessible from the
main menu.