Try to reduce rolling shutter effect
The rolling shutter effect occurs when there is a significant delay
between reading different points on the light sensor array. This becomes
an issue if there is too much delay between reading the upper and lower
sensors in a pair, specifically when the screen is changing color.

It ain't much, but it's honest work.